Added Script for creating Tower

+++ Assets/Builder/Tower.cs
@@ -1,5 +1,5 @@
 public class Tower
 {
     public int Level { get; set; }
-    public float Accuracy { get; set; }
+    public double Accuracy { get; set; }
 }
